Move shared modules into module directory
