+++ src/mopie_data/video_metadata_s3_manager.py
@@ -1,6 +1,6 @@
 import boto3
 import json
-from models import VideoMetadata
+from mopie_common.models import VideoMetadata
 from itertools import chain
 
 s3 = boto3.client('s3')
